demo-example/requirements.txt
pandas>=2.2.3
streamlit>=1.45.1
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
plotly>=5.15.0 